+++ frontend/mja-ws/libs/aufgabensammlungen/data/src/lib/aufgabensammlungen.actions.ts
@@ -9,7 +9,7 @@ export const aufgabensammlungenActions = createActionGroup({
         'FIND_AUFGABENSAMMLUNGEN': props<{aufgabensammlungenSuchparameter: AufgabensammlungenSuchparameter, pageDefinition: PageDefinition}>(),
         'AUFGABENSAMMLUNGEN_FOUND': props<{treffer: AufgabensammlungenTreffer}>(),
         'AUFGABENSAMMLUNGEN_SELECT_PAGE': props<{pageDefinition: PageDefinition}>(),
-        'SELECT_AUFGABENSAMMLUNG': props<{ aufgabensammlung: AufgabensammlungTrefferItem }>(),
+        'SELECT_AUFGABENSAMMLUNG': props<{ aufgabensammlungId: string }>(),
         'UNSELECT_AUFGABENSAMMLUNG': emptyProps(),
         'AUFGABENSAMMLUNGDETAILS_LOADED': props<{aufgabensammlungDetails: AufgabensammlungDetails, navigateTo: string}>(),
         'GENERIERE_ARBEITSBLATT': props<{aufgabensammlungID: string, font: FontName, schriftgroesse: Schriftgroesse, layoutAntwortvorschlaege: LaTeXLayoutAntwortvorschlaege}>(),